+++ changelog.md
@@ -41,6 +41,7 @@ This version is not yet released. If you are reading this on the website, then t
 - Functions are now analyzed for purity
   - A `|0.1` binding will be automatically evaluated at compile time if it is pure
 - Language Server
+  - Add signature inlay hints
   - Add macro expansion as a code action
   - Add completions for shadowable constants
   - Add completions for module items when the module reference is partially typed